fix for measuring Units

+++ SET3-Backend/Models/ProductShopIntertable.cs
@@ -8,9 +8,9 @@ namespace SET3_Backend.Models
         public int Id { get; set; }
         public int ShopId { get; set; }
         public int ProductId { get; set; }
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
 
-        public ProductShopIntertable(int shopId, int productId, int quantity)
+        public ProductShopIntertable(int shopId, int productId, double quantity)
         {
             ShopId = shopId;
             ProductId = productId;
